backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
lancedb==0.5.4
redis==5.0.1
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
numpy==1.26.0
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
openai==1.3.7  # For Azure OpenAI
google-generativeai==0.3.2
boto3==1.34.0
anthropic==0.7.8
clip-by-openai==1.1.0  # Fallback for embeddings
torch==2.1.0
transformers==4.35.2
pyarrow==14.0.1
pandas==2.1.3